Put import/export into its own component
