+++ src/components/theme/theme-editor.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import { ThemeVariable } from '$lib/theme/theme-variable.svelte';
+	import ImportExport from '../import-export.svelte';
 
 	const backgroundDefault = '#8c8cff';
 	const textDefault = '#ffff8c';
@@ -34,73 +35,6 @@
 			})
 		);
 	});
-
-	function exportData() {
-		const localStorageContents = Object.fromEntries(
-			[
-				'color-theme',
-				'custom-colors',
-				'filled-examples',
-				'link-storage_read',
-				'link-storage_unread'
-			].map((key) => {
-				const value = localStorage.getItem(key);
-				return [key, value];
-			})
-		);
-
-		const blob = new Blob([JSON.stringify(localStorageContents, null, 2)], {
-			type: 'application/json'
-		});
-
-		const objectURL = URL.createObjectURL(blob);
-		try {
-			const a = document.createElement('a');
-			a.download = 'reading-list-data.json';
-			a.href = objectURL;
-			a.click();
-		} finally {
-			URL.revokeObjectURL(objectURL);
-		}
-	}
-
-	async function importData() {
-		const input = document.createElement('input') as HTMLInputElement;
-		input.type = 'file';
-
-		const filePromise = new Promise((resolve) => {
-			input.addEventListener(
-				'change',
-				() => {
-					resolve(input.files?.[0]);
-				},
-				{ once: true }
-			);
-		});
-		input.click();
-
-		const file = (await filePromise) as File | undefined;
-
-		if (!file) {
-			throw new Error('No file contents :(');
-		}
-
-		const fileContents = await file.text();
-		const localStorageContents = JSON.parse(fileContents);
-		for (const [key, value] of Object.entries(localStorageContents)) {
-			if (typeof key !== 'string' || typeof value !== 'string') {
-				throw new Error('Bad data');
-			}
-		}
-
-		localStorage.clear();
-
-		for (const [key, value] of Object.entries(localStorageContents)) {
-			localStorage.setItem(key, value as string);
-		}
-
-		location.reload();
-	}
 </script>
 
 <div class="theme-editor">
@@ -145,12 +79,5 @@
 		</li>
 	</menu>
 
-	<menu class="buttons">
-		<li>
-			<button type="button" onclick={importData}>Import data</button>
-		</li>
-		<li>
-			<button type="button" onclick={exportData}>Export data</button>
-		</li>
-	</menu>
+	<ImportExport />
 </div>
